Agrego Ejercicios 1, 2 y 3 de la Práctica 2
